api/python/requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
django==4.2.2
django-cors-headers==4.1.0
django-filter==23.2
djangorestframework==3.14.0
markdown==3.4.3
pytz==2023.3
sqlparse==0.4.4 ; python_version >= '3.5'
tzdata==2023.3 ; sys_platform == 'win32'
